Remove comment and adjust flexbox properties.
Remove comment for button hover.
Adjust properties for flexbox in the header section so that the image would wrap below the text and button on smaller browser windows
Create a max width and add flex-wrap to the signup section so the button would wrap below the text on smaller browser windows.